pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "observascore"
version = "0.1.0"
description = "Observability & SRE Maturity Assessment Tool"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "ObservaScore Contributors" }]
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "click>=8.1",
    "rich>=13.7",
    "pydantic>=2.5",
]

[project.scripts]
observascore = "observascore.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["observascore*"]

[tool.setuptools.package-data]
"observascore.rules" = ["packs/*.yaml"]
"observascore.report" = ["templates/*.html", "templates/*.css"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
